test: merge exception tests to a controller test file and separate package
